+++ NEWS.md
@@ -4,6 +4,7 @@
   results to R. Previously, there were multiple data copies between R and C++
   that added overhead to the computations.
 * The previous versions returned MX by default, now it has to be specified.
+* Adds code to extract the fixed effects with `felm` objects.
 
 # capybara 0.5.2